# courseGenie_backEnd/docs/apidocs/ui/openapi.json
{
  "openapi": "3.0.3",
  "info" : {
    "description" : "<h1>course-genie-backend<\/h1><p>Demo project for Spring Boot<\/p>",
    "version" : "0.0.1-SNAPSHOT",
    "title" : "course-genie-backend"
  },
  "servers" : [
    {
      "url" : "/"
    }
  ],
...
  "components" : {
    "schemas" : {
...
      "AssessmentDTO" : {
"type" : "object",
"title" : "AssessmentDTO",
"required" : ["maxPoints", "sectionId", "assessmentId", "week"],
"properties" : {
  "maxPoints" : {
    "readOnly" : false,
"description" : "",
"type" : "integer",
"format" : "int32"
  },
  "sectionId" : {
    "readOnly" : false,
"description" : "",
"type" : "integer",
"format" : "int64"
  },
  "shortName" : {
    "readOnly" : false,
"description" : "",
"type" : "string"
  },
  "clos" : {
    "readOnly" : false,
"description" : "",
"type" : "array",
"items" : {
  "$ref" : "#/components/schemas/CLODTO"
}
  },
  "name" : {
    "readOnly" : false,
"description" : "",
"type" : "string"
  },
  "category" : {
    "readOnly" : false,
"description" : "",
"type" : "string"
  },
  "assessmentId" : {
    "readOnly" : false,
"description" : "",
"type" : "integer",
"format" : "int64"
  },
  "week" : {
    "readOnly" : false,
"description" : "",
"type" : "integer",
"format" : "int32"
  }
},
"example" : {
  "maxPoints" : 12345,
  "sectionId" : 12345,
  "shortName" : "...",
  "clos" : [ {
    "courseId" : 12345,
    "description" : "...",
    "name" : "...",
    "cloId" : 12345
  }, {
    "courseId" : 12345,
    "description" : "...",
    "name" : "...",
    "cloId" : 12345
  } ],
  "name" : "...",
  "category" : "...",
  "assessmentId" : 12345,
  "week" : 12345
},
"description" : ""
      }
...
      "AuthQuery" : {
"type" : "object",
"title" : "AuthQuery",
"properties" : {
  "password" : {
    "readOnly" : false,
"description" : "",
"type" : "string"
  },
  "username" : {
    "readOnly" : false,
"description" : "",
"type" : "string"
  }
},
"example" : {
  "password" : "...",
  "username" : "..."
},
"description" : ""
      }
...
      "CLODTO" : {
"type" : "object",
"title" : "CLODTO",
"required" : ["courseId", "cloId"],
"properties" : {
  "courseId" : {
    "readOnly" : false,
"description" : "",
"type" : "integer",
"format" : "int64"
  },
  "description" : {
    "readOnly" : false,
"description" : "",
"type" : "string"
  },
  "name" : {
    "readOnly" : false,
"description" : "",
"type" : "string"
  },
  "cloId" : {
    "readOnly" : false,
"description" : "",
"type" : "integer",
"format" : "int64"
  }
},
"example" : {
  "courseId" : 12345,
  "description" : "...",
  "name" : "...",
  "cloId" : 12345
},
"description" : ""
      }
...
      "CourseDTO" : {
"type" : "object",
"title" : "CourseDTO",
"required" : ["courseId"],
"properties" : {
  "courseId" : {
    "readOnly" : false,
"description" : "",
"type" : "integer",
"format" : "int64"
  },
  "code" : {
    "readOnly" : false,
"description" : "",
"type" : "string"
  },
  "credits" : {
    "readOnly" : false,
"description" : "",
"type" : "string"
  },
  "sections" : {
    "readOnly" : false,
"description" : "",
"type" : "array",
"items" : {
  "$ref" : "#/components/schemas/SectionDTO"
}
  },
  "description" : {
    "readOnly" : false,
"description" : "",
"type" : "string"
  },
  "name" : {
    "readOnly" : false,
"description" : "",
"type" : "string"
  },
  "clos" : {
    "readOnly" : false,
"description" : "",
"type" : "array",
"items" : {
  "$ref" : "#/components/schemas/CLODTO"
}
  }
},
"example" : {
  "courseId" : 12345,
  "code" : "...",
  "credits" : "...",
  "sections" : [ {
    "teachingMethodology" : "...",
    "sectionId" : 12345,
    "professorId" : 12345,
    "code" : "...",
    "term" : "...",
    "configured" : true,
    "assessments" : [ {
      "maxPoints" : 12345,
      "sectionId" : 12345,
      "shortName" : "...",
      "clos" : [ { }, { } ],
      "name" : "...",
      "category" : "...",
      "assessmentId" : 12345,
      "week" : 12345
    }, {
      "maxPoints" : 12345,
      "sectionId" : 12345,
      "shortName" : "...",
      "clos" : [ { }, { } ],
      "name" : "...",
      "category" : "...",
      "assessmentId" : 12345,
      "week" : 12345
    } ],
    "class_number" : "...",
    "courseId" : 12345
  }, {
    "teachingMethodology" : "...",
    "sectionId" : 12345,
    "professorId" : 12345,
    "code" : "...",
    "term" : "...",
    "configured" : true,
    "assessments" : [ {
      "maxPoints" : 12345,
      "sectionId" : 12345,
      "shortName" : "...",
      "clos" : [ { }, { } ],
      "name" : "...",
      "category" : "...",
      "assessmentId" : 12345,
      "week" : 12345
    }, {
      "maxPoints" : 12345,
      "sectionId" : 12345,
      "shortName" : "...",
      "clos" : [ { }, { } ],
      "name" : "...",
      "category" : "...",
      "assessmentId" : 12345,
      "week" : 12345
    } ],
    "class_number" : "...",
    "courseId" : 12345
  } ],
  "description" : "...",
  "name" : "...",
  "clos" : [ {
    "courseId" : 12345,
    "description" : "...",
    "name" : "...",
    "cloId" : 12345
  }, {
    "courseId" : 12345,
    "description" : "...",
    "name" : "...",
    "cloId" : 12345
  } ]
},
"description" : ""
      }
...
      "GradeDTO" : {
"type" : "object",
"title" : "GradeDTO",
"required" : ["score", "gradeId", "assessmentId"],
"properties" : {
  "studentId" : {
    "readOnly" : false,
"description" : "",
"type" : "string"
  },
  "score" : {
    "readOnly" : false,
"description" : "",
"type" : "number",
"format" : "double"
  },
  "gradeId" : {
    "readOnly" : false,
"description" : "",
"type" : "integer",
"format" : "int64"
  },
  "assessmentId" : {
    "readOnly" : false,
"description" : "",
"type" : "integer",
"format" : "int64"
  }
},
"example" : {
  "studentId" : "...",
  "score" : 12345.0,
  "gradeId" : 12345,
  "assessmentId" : 12345
},
"description" : ""
      }
...
      "SectionDTO" : {
"type" : "object",
"title" : "SectionDTO",
"required" : ["sectionId", "professorId", "configured", "courseId"],
"properties" : {
  "teachingMethodology" : {
    "readOnly" : false,
"description" : "",
"type" : "string"
  },
  "sectionId" : {
    "readOnly" : false,
"description" : "",
"type" : "integer",
"format" : "int64"
  },
  "professorId" : {
    "readOnly" : false,
"description" : "",
"type" : "integer",
"format" : "int64"
  },
  "code" : {
    "readOnly" : false,
"description" : "",
"type" : "string"
  },
  "term" : {
    "readOnly" : false,
"description" : "",
"type" : "string"
  },
  "configured" : {
    "readOnly" : false,
"description" : "",
"type" : "boolean"
  },
  "assessments" : {
    "readOnly" : false,
"description" : "",
"type" : "array",
"items" : {
  "$ref" : "#/components/schemas/AssessmentDTO"
}
  },
  "class_number" : {
    "readOnly" : false,
"description" : "",
"type" : "string"
  },
  "courseId" : {
    "readOnly" : false,
"description" : "",
"type" : "integer",
"format" : "int64"
  }
},
"example" : {
  "teachingMethodology" : "...",
  "sectionId" : 12345,
  "professorId" : 12345,
  "code" : "...",
  "term" : "...",
  "configured" : true,
  "assessments" : [ {
    "maxPoints" : 12345,
    "sectionId" : 12345,
    "shortName" : "...",
    "clos" : [ {
      "courseId" : 12345,
      "description" : "...",
      "name" : "...",
      "cloId" : 12345
    }, {
      "courseId" : 12345,
      "description" : "...",
      "name" : "...",
      "cloId" : 12345
    } ],
    "name" : "...",
    "category" : "...",
    "assessmentId" : 12345,
    "week" : 12345
  }, {
    "maxPoints" : 12345,
    "sectionId" : 12345,
    "shortName" : "...",
    "clos" : [ {
      "courseId" : 12345,
      "description" : "...",
      "name" : "...",
      "cloId" : 12345
    }, {
      "courseId" : 12345,
      "description" : "...",
      "name" : "...",
      "cloId" : 12345
    } ],
    "name" : "...",
    "category" : "...",
    "assessmentId" : 12345,
    "week" : 12345
  } ],
  "class_number" : "...",
  "courseId" : 12345
},
"description" : ""
      }
...
      "xml_ns0_CLODTO" : {
"type" : "object",
"title" : "CLODTO",
"required" : ["cloId", "courseId"],
"allOf" : [
  {
"properties" : {
  "cloId" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "integer",
"format" : "int64"
  },
  "courseId" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "integer",
"format" : "int64"
  },
  "description" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "string"
  },
  "name" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "string"
  }
}
  }
],
"description" : ""
      }
...
      "xml_ns0_courseDTO" : {
"type" : "object",
"title" : "courseDTO",
"required" : ["courseId"],
"allOf" : [
  {
"properties" : {
  "clos" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "array",
"items" : {
  "$ref" : "#/components/schemas/xml_ns0_CLODTO"
}
  },
  "code" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "string"
  },
  "courseId" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "integer",
"format" : "int64"
  },
  "credits" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "string"
  },
  "description" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "string"
  },
  "name" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "string"
  },
  "sections" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "array",
"items" : {
  "$ref" : "#/components/schemas/xml_ns0_sectionDTO"
}
  }
}
  }
],
"description" : ""
      }
...
      "xml_ns0_gradeDTO" : {
"type" : "object",
"title" : "gradeDTO",
"required" : ["assessmentId", "gradeId", "score"],
"allOf" : [
  {
"properties" : {
  "assessmentId" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "integer",
"format" : "int64"
  },
  "gradeId" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "integer",
"format" : "int64"
  },
  "score" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "number",
"format" : "double"
  },
  "studentId" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "string"
  }
}
  }
],
"description" : ""
      }
...
      "xml_ns0_sectionDTO" : {
"type" : "object",
"title" : "sectionDTO",
"required" : ["configured", "courseId", "professorId", "sectionId"],
"allOf" : [
  {
"properties" : {
  "assessments" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "array",
"items" : {
  "$ref" : "#/components/schemas/xml_ns0_assessmentDTO"
}
  },
  "class_number" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "string"
  },
  "code" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "string"
  },
  "configured" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "boolean"
  },
  "courseId" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "integer",
"format" : "int64"
  },
  "professorId" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "integer",
"format" : "int64"
  },
  "sectionId" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "integer",
"format" : "int64"
  },
  "teachingMethodology" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "string"
  },
  "term" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "string"
  }
}
  }
],
"description" : ""
      }
...
      "xml_ns0_authQuery" : {
"type" : "object",
"title" : "authQuery",
"allOf" : [
  {
"properties" : {
  "password" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "string"
  },
  "username" : {
    "xml" : {
    },
    "readOnly" : false,
"description" : "",
"type" : "string"
  }
}
  }
],
"description" : ""
      }
...
    }
  },
  "paths": {
    "\/api/admin/roles/assign" : {
      "post" : {
        "tags" : [ "RoleManagementController" ],
        "operationId" : "assignRole",
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
"description" : "",
"type" : "string"
                }
              }, 
              "application/xml" : {
                "schema" : {
"description" : "",
"type" : "string"
                }
              }
            },
            "description" : ""
          }
        },
        "parameters" : [
          {
            "name" : "role",
            "in" : "query",
            "required" : true,
            "schema" : {
              "type" : "string"
            },
            "description" : ""
          },
          {
            "name" : "username",
            "in" : "query",
            "required" : true,
            "schema" : {
              "type" : "string"
            },
            "description" : ""
          }
        ]
      }
    },
    "\/api/admin/roles/remove" : {
      "post" : {
        "tags" : [ "RoleManagementController" ],
        "operationId" : "removeRole",
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
"description" : "",
"type" : "string"
                }
              }, 
              "application/xml" : {
                "schema" : {
"description" : "",
"type" : "string"
                }
              }
            },
            "description" : ""
          }
        },
        "parameters" : [
          {
            "name" : "role",
            "in" : "query",
            "required" : true,
            "schema" : {
              "type" : "string"
            },
            "description" : ""
          },
          {
            "name" : "username",
            "in" : "query",
            "required" : true,
            "schema" : {
              "type" : "string"
            },
            "description" : ""
          }
        ]
      }
    },
...
    "\/api/auth/authenticate" : {
      "post" : {
        "tags" : [ "JWTAuthenticationController" ],
        "operationId" : "authenticate",
        "requestBody" : {
          "required" : false,
          "content" : {
            "application/json" : {
              "examples" : {
                "example" : {
                  "value" : "{\n  \"password\" : \"...\",\n  \"username\" : \"...\"\n}"
                }
              },
              "schema" : {
"$ref" : "#/components/schemas/AuthQuery",
"description" : ""
              }
            }, 
            "application/xml" : {
              "examples" : {
                "example" : {
                  "value" : "<a-----\u003E\n  <password>...<\/password>\n  <username>...<\/username>\n<\/a-----\u003E\n"
                }
              },
              "schema" : {
"$ref" : "#/components/schemas/xml_ns0_authQuery",
"description" : ""
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
"description" : "",
"type" : "string"
                }
              }, 
              "application/xml" : {
                "schema" : {
"description" : "",
"type" : "string"
                }
              }
            },
            "description" : ""
          }
        },
        "parameters" : [
        ]
      }
    },
...
    "\/api/courses/{courseCode}/section/{sectionCode}" : {
      "get" : {
        "tags" : [ "CourseController" ],
        "operationId" : "getCourseById",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "example" : {
                    "value" : "{\n  \"courseId\" : 12345,\n  \"code\" : \"...\",\n  \"credits\" : \"...\",\n  \"sections\" : [ {\n    \"teachingMethodology\" : \"...\",\n    \"sectionId\" : 12345,\n    \"professorId\" : 12345,\n    \"code\" : \"...\",\n    \"term\" : \"...\",\n    \"configured\" : true,\n    \"assessments\" : [ {\n      \"maxPoints\" : 12345,\n      \"sectionId\" : 12345,\n      \"shortName\" : \"...\",\n      \"clos\" : [ { }, { } ],\n      \"name\" : \"...\",\n      \"category\" : \"...\",\n      \"assessmentId\" : 12345,\n      \"week\" : 12345\n    }, {\n      \"maxPoints\" : 12345,\n      \"sectionId\" : 12345,\n      \"shortName\" : \"...\",\n      \"clos\" : [ { }, { } ],\n      \"name\" : \"...\",\n      \"category\" : \"...\",\n      \"assessmentId\" : 12345,\n      \"week\" : 12345\n    } ],\n    \"class_number\" : \"...\",\n    \"courseId\" : 12345\n  }, {\n    \"teachingMethodology\" : \"...\",\n    \"sectionId\" : 12345,\n    \"professorId\" : 12345,\n    \"code\" : \"...\",\n    \"term\" : \"...\",\n    \"configured\" : true,\n    \"assessments\" : [ {\n      \"maxPoints\" : 12345,\n      \"sectionId\" : 12345,\n      \"shortName\" : \"...\",\n      \"clos\" : [ { }, { } ],\n      \"name\" : \"...\",\n      \"category\" : \"...\",\n      \"assessmentId\" : 12345,\n      \"week\" : 12345\n    }, {\n      \"maxPoints\" : 12345,\n      \"sectionId\" : 12345,\n      \"shortName\" : \"...\",\n      \"clos\" : [ { }, { } ],\n      \"name\" : \"...\",\n      \"category\" : \"...\",\n      \"assessmentId\" : 12345,\n      \"week\" : 12345\n    } ],\n    \"class_number\" : \"...\",\n    \"courseId\" : 12345\n  } ],\n  \"description\" : \"...\",\n  \"name\" : \"...\",\n  \"clos\" : [ {\n    \"courseId\" : 12345,\n    \"description\" : \"...\",\n    \"name\" : \"...\",\n    \"cloId\" : 12345\n  }, {\n    \"courseId\" : 12345,\n    \"description\" : \"...\",\n    \"name\" : \"...\",\n    \"cloId\" : 12345\n  } ]\n}"
                  }
                },
                "schema" : {
"$ref" : "#/components/schemas/CourseDTO",
"description" : ""
                }
              }, 
              "application/xml" : {
                "examples" : {
                  "example" : {
                    "value" : "<c-----\u003E\n  <credits>...<\/credits>\n  <code>...<\/code>\n  <name>...<\/name>\n  <clos>\n    <cloId>...<\/cloId>\n    <description>...<\/description>\n    <name>...<\/name>\n    <courseId>...<\/courseId>\n  <\/clos>\n  <courseId>...<\/courseId>\n  <sections>\n    <sectionId>...<\/sectionId>\n    <teachingMethodology>...<\/teachingMethodology>\n    <class_number>...<\/class_number>\n    <code>...<\/code>\n    <assessments>\n      <category>...<\/category>\n      <sectionId>...<\/sectionId>\n      <shortName>...<\/shortName>\n      <name>...<\/name>\n      <maxPoints>...<\/maxPoints>\n      <assessmentId>...<\/assessmentId>\n      <clos/>\n      <week>...<\/week>\n    <\/assessments>\n    <professorId>...<\/professorId>\n    <courseId>...<\/courseId>\n    <configured>...<\/configured>\n    <term>...<\/term>\n  <\/sections>\n  <description>...<\/description>\n<\/c-----\u003E\n"
                  }
                },
                "schema" : {
"$ref" : "#/components/schemas/xml_ns0_courseDTO",
"description" : ""
                }
              }
            },
            "description" : ""
          }
        },
        "parameters" : [
          {
            "name" : "courseCode",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            },
            "description" : ""
          },
          {
            "name" : "sectionCode",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "string"
            },
            "description" : ""
          }
        ]
      }
    },
...
    "\/api/grades" : {
      "get" : {
        "tags" : [ "GradeController" ],
        "operationId" : "getAllGrades",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "example" : {
                    "value" : "[ {\n  \"studentId\" : \"...\",\n  \"score\" : 12345.0,\n  \"gradeId\" : 12345,\n  \"assessmentId\" : 12345\n} ]"
                  }
                },
                "schema" : {
"description" : "",
"type" : "array",
"items" : {
  "$ref" : "#/components/schemas/GradeDTO"
}
                }
              }, 
              "application/xml" : {
                "examples" : {
                  "example" : {
                    "value" : "<list>\n  <g-----\u003E\n    <gradeId>...<\/gradeId>\n    <studentId>...<\/studentId>\n    <assessmentId>...<\/assessmentId>\n    <score>...<\/score>\n  <\/g-----\u003E\n<\/list>\n"
                  }
                },
                "schema" : {
"description" : "",
"type" : "array",
"items" : {
  "$ref" : "#/components/schemas/xml_ns0_gradeDTO"
}
                }
              }
            },
            "description" : ""
          }
        },
        "parameters" : [
        ]
      },
      "post" : {
        "tags" : [ "GradeController" ],
        "operationId" : "createGrade",
        "requestBody" : {
          "required" : false,
          "content" : {
            "application/json" : {
              "examples" : {
                "example" : {
                  "value" : "[ {\n  \"studentId\" : \"...\",\n  \"score\" : 12345.0,\n  \"gradeId\" : 12345,\n  \"assessmentId\" : 12345\n} ]"
                }
              },
              "schema" : {
"description" : "",
"type" : "array",
"items" : {
  "$ref" : "#/components/schemas/GradeDTO"
}
              }
            }, 
            "application/xml" : {
              "examples" : {
                "example" : {
                  "value" : "<list>\n  <g-----\u003E\n    <gradeId>...<\/gradeId>\n    <studentId>...<\/studentId>\n    <assessmentId>...<\/assessmentId>\n    <score>...<\/score>\n  <\/g-----\u003E\n<\/list>\n"
                }
              },
              "schema" : {
"description" : "",
"type" : "array",
"items" : {
  "$ref" : "#/components/schemas/xml_ns0_gradeDTO"
}
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
"description" : "",
"type" : "boolean"
                }
              }, 
              "application/xml" : {
                "schema" : {
"description" : "",
"type" : "boolean"
                }
              }
            },
            "description" : ""
          }
        },
        "parameters" : [
        ]
      },
      "put" : {
        "tags" : [ "GradeController" ],
        "operationId" : "updateGrade",
        "requestBody" : {
          "required" : false,
          "content" : {
            "application/json" : {
              "examples" : {
                "example" : {
                  "value" : "{\n  \"studentId\" : \"...\",\n  \"score\" : 12345.0,\n  \"gradeId\" : 12345,\n  \"assessmentId\" : 12345\n}"
                }
              },
              "schema" : {
"$ref" : "#/components/schemas/GradeDTO",
"description" : ""
              }
            }, 
            "application/xml" : {
              "examples" : {
                "example" : {
                  "value" : "<g-----\u003E\n  <gradeId>...<\/gradeId>\n  <studentId>...<\/studentId>\n  <assessmentId>...<\/assessmentId>\n  <score>...<\/score>\n<\/g-----\u003E\n"
                }
              },
              "schema" : {
"$ref" : "#/components/schemas/xml_ns0_gradeDTO",
"description" : ""
              }
            }
          }
        },
        "responses" : {
          "204" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "example" : {
                    "value" : "{\n  \"studentId\" : \"...\",\n  \"score\" : 12345.0,\n  \"gradeId\" : 12345,\n  \"assessmentId\" : 12345\n}"
                  }
                },
                "schema" : {
"$ref" : "#/components/schemas/GradeDTO",
"description" : ""
                }
              }, 
              "application/xml" : {
                "examples" : {
                  "example" : {
                    "value" : "<g-----\u003E\n  <gradeId>...<\/gradeId>\n  <studentId>...<\/studentId>\n  <assessmentId>...<\/assessmentId>\n  <score>...<\/score>\n<\/g-----\u003E\n"
                  }
                },
                "schema" : {
"$ref" : "#/components/schemas/xml_ns0_gradeDTO",
"description" : ""
                }
              }
            },
            "description" : ""
          }
        },
        "parameters" : [
        ]
      }
    },
    "\/api/grades/{gradeId}" : {
      "delete" : {
        "tags" : [ "GradeController" ],
        "operationId" : "deleteGrade",
        "responses" : {
          "204" : {
            "content" : {
              "application/json" : {
                "schema" : {
"description" : "",
"type" : "boolean"
                }
              }, 
              "application/xml" : {
                "schema" : {
"description" : "",
"type" : "boolean"
                }
              }
            },
            "description" : ""
          }
        },
        "parameters" : [
          {
            "name" : "gradeId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "number"
            },
            "description" : ""
          }
        ]
      },
      "get" : {
        "tags" : [ "GradeController" ],
        "operationId" : "getGradeById",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "examples" : {
                  "example" : {
                    "value" : "{\n  \"studentId\" : \"...\",\n  \"score\" : 12345.0,\n  \"gradeId\" : 12345,\n  \"assessmentId\" : 12345\n}"
                  }
                },
                "schema" : {
"$ref" : "#/components/schemas/GradeDTO",
"description" : ""
                }
              }, 
              "application/xml" : {
                "examples" : {
                  "example" : {
                    "value" : "<g-----\u003E\n  <gradeId>...<\/gradeId>\n  <studentId>...<\/studentId>\n  <assessmentId>...<\/assessmentId>\n  <score>...<\/score>\n<\/g-----\u003E\n"
                  }
                },
                "schema" : {
"$ref" : "#/components/schemas/xml_ns0_gradeDTO",
"description" : ""
                }
              }
            },
            "description" : ""
          }
        },
        "parameters" : [
          {
            "name" : "gradeId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "number"
            },
            "description" : ""
          }
        ]
      }
    },
    "\/api/sections/{sectionId}" : {
      "put" : {
        "tags" : [ "SectionController" ],
        "operationId" : "saveConfiguration",
        "responses" : {
          "204" : {
            "content" : {
              "application/json" : {
                "schema" : {
"description" : "",
"type" : "boolean"
                }
              }, 
              "application/xml" : {
                "schema" : {
"description" : "",
"type" : "boolean"
                }
              }
            },
            "description" : ""
          }
        },
        "parameters" : [
          {
            "name" : "sectionId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "number"
            },
            "description" : ""
          }
        ]
      }
    },
    "\/api/sections/{sectionId}/teaching-methodology" : {
      "get" : {
        "tags" : [ "SectionController" ],
        "operationId" : "getTeachingMethodology",
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
"description" : "",
"type" : "string"
                }
              }, 
              "application/xml" : {
                "schema" : {
"description" : "",
"type" : "string"
                }
              }
            },
            "description" : ""
          }
        },
        "parameters" : [
          {
            "name" : "sectionId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "number"
            },
            "description" : ""
          }
        ]
      },
      "post" : {
        "tags" : [ "SectionController" ],
        "operationId" : "setTeachingMethodology",
        "requestBody" : {
          "required" : false,
          "content" : {
            "application/json" : {
              "schema" : {
"description" : "",
"type" : "string"
              }
            }, 
            "application/xml" : {
              "schema" : {
"description" : "",
"type" : "string"
              }
            }
          }
        },
        "responses" : {
          "201" : {
            "content" : {
            },
            "description" : "Success"
          }
        },
        "parameters" : [
          {
            "name" : "sectionId",
            "in" : "path",
            "required" : true,
            "schema" : {
              "type" : "number"
            },
            "description" : ""
          }
        ]
      }
    },
...
  }
}